--- SimPEG/Utils/printinfo.py
@@ -1,111 +1,0 @@
-try:  # scooby is a soft dependencies for discretize
-    import scooby
-    from scooby import Report
-except ImportError:
-    scooby = False
-
-    class Report:
-        pass
-
-
-def requires(modules):
-    """Decorator to wrap functions with soft dependencies.
-
-    This function was inspired by the `requires` function of pysal,
-    which is released under the 'BSD 3-Clause "New" or "Revised" License'.
-
-    https://github.com/pysal/pysal/blob/master/pysal/lib/common.py
-
-    Parameters
-    ----------
-    modules : dict
-        Dictionary containing soft dependencies, e.g.,
-        {'matplotlib': matplotlib}.
-
-    Returns
-    -------
-    decorated_function : function
-        Original function if all soft dependencies are met, otherwise
-        it returns an empty function which prints why it is not running.
-
-    """
-
-    # Check the required modules, add missing ones in the list `missing`.
-    missing = []
-    for key, item in modules.items():
-        if item is False:
-            missing.append(key)
-
-    def decorated_function(function):
-        """Wrap function."""
-        if not missing:
-            return function
-        else:
-            def passer(*args, **kwargs):
-                print(('Missing dependencies: {d}.'.format(d=missing)))
-                print(('Not running `{}`.'.format(function.__name__)))
-            return passer
-
-    return decorated_function
-
-
-@requires({'scooby': scooby})
-class Report(Report):
-    """Print date, time, and version information.
-
-    Use scooby to print date, time, and package version information in any
-    environment (Jupyter notebook, IPython console, Python console, QT
-    console), either as html-table (notebook) or as plain text (anywhere).
-
-    Always shown are the OS, number of CPU(s), ``numpy``, ``scipy``,
-    ``SimPEG``, ``cython``, ``properties``, ``vectormath``, ``discretize``,
-    ``pymatsolver``, ``sys.version``, and time/date.
-
-    Additionally shown are, if they can be imported, ``IPython``,
-    ``matplotlib``, and ``ipywidgets``. It also shows MKL information, if
-    available.
-
-    All modules provided in ``add_pckg`` are also shown.
-
-
-    Parameters
-    ----------
-    add_pckg : packages, optional
-        Package or list of packages to add to output information (must be
-        imported beforehand).
-
-    ncol : int, optional
-        Number of package-columns in html table (no effect in text-version);
-        Defaults to 3.
-
-    text_width : int, optional
-        The text width for non-HTML display modes
-
-    sort : bool, optional
-        Sort the packages when the report is shown
-
-
-    Examples
-    --------
-
-    >>> import pytest
-    >>> import dateutil
-    >>> from SimPEG import Report
-    >>> Report()                            # Default values
-    >>> Report(pytest)                      # Provide additional package
-    >>> Report([pytest, dateutil], ncol=5)  # Define nr of columns
-
-    """
-
-    def __init__(self, add_pckg=None, ncol=3, text_width=80, sort=False):
-        """Initiate a scooby.Report instance."""
-
-        # Mandatory packages.
-        core = ['SimPEG', 'discretize', 'pymatsolver', 'vectormath',
-                'properties', 'numpy', 'scipy', 'cython']
-
-        # Optional packages.
-        optional = ['IPython', 'matplotlib', 'ipywidgets']
-
-        super().__init__(additional=add_pckg, core=core, optional=optional,
-                         ncol=ncol, text_width=text_width, sort=sort)
