+++ docs/content/api_core/api_Utils.rst
@@ -5,13 +5,6 @@ Utils
     :members:
     :undoc-members:
 
-Version
-=======
-
-.. automodule:: SimPEG.Utils.printinfo
-    :members:
-    :undoc-members:
-
 
 Matrix Utilities
 ================
